backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
sendgrid==6.11.0
reportlab==4.0.0
PyJWT==2.8.0
werkzeug==2.3.7